[luci] Relocate CloneNode class
This will relocate CloneNode class to it's own file and into CircleConst
to reuse methods.

ONE-DCO-1.0-Signed-off-by: SaeHie Park <saehie.park@gmail.com>
