+++ compiler/luci/service/src/CircleNodeClone.cpp
@@ -16,151 +16,11 @@
 
 #include "luci/Service/CircleNodeClone.h"
 
-#include <luci/IR/CircleNodeVisitor.h>
+#include "CircleCloneNode.h"
 
 #include <oops/UserExn.h>
 
 #include <cassert>
-
-namespace
-{
-
-class CloneNode final : public luci::CircleNodeVisitor<luci::CircleNode *>
-{
-public:
-  CloneNode(loco::Graph *graph) : _graph(graph){};
-
-public:
-  luci::CircleNode *visit(const luci::CircleAdd *) final;
-  luci::CircleNode *visit(const luci::CircleConst *) final;
-  luci::CircleNode *visit(const luci::CircleDiv *) final;
-  luci::CircleNode *visit(const luci::CircleMean *) final;
-  luci::CircleNode *visit(const luci::CircleMul *) final;
-  luci::CircleNode *visit(const luci::CirclePow *) final;
-  luci::CircleNode *visit(const luci::CircleRsqrt *) final;
-  luci::CircleNode *visit(const luci::CircleSqrt *) final;
-  luci::CircleNode *visit(const luci::CircleSquaredDifference *) final;
-  luci::CircleNode *visit(const luci::CircleSub *) final;
-  // TODO add all nodes
-
-  // NOTE CircleNodeVisitor will throw if not supported here
-
-protected:
-  loco::Graph *_graph = nullptr;
-};
-
-luci::CircleNode *CloneNode::visit(const luci::CircleAdd *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleAdd>();
-  cloned->fusedActivationFunction(node->fusedActivationFunction());
-  return cloned;
-}
-
-template <loco::DataType T>
-void copy_values(const luci::CircleConst *node, luci::CircleConst *cloned)
-{
-  assert(T == node->dtype());
-  assert(T == cloned->dtype());
-
-  const auto size = node->size<T>();
-  cloned->size<T>(size);
-  for (uint32_t i = 0; i < size; i++)
-    cloned->at<T>(i) = node->at<T>(i);
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleConst *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleConst>();
-
-  cloned->dtype(node->dtype());
-  cloned->rank(node->rank());
-
-  switch (node->dtype())
-  {
-    case loco::DataType::FLOAT32:
-      copy_values<loco::DataType::FLOAT32>(node, cloned);
-      break;
-
-    case loco::DataType::U8:
-      copy_values<loco::DataType::U8>(node, cloned);
-      break;
-
-    case loco::DataType::S8:
-      copy_values<loco::DataType::S8>(node, cloned);
-      break;
-
-    case loco::DataType::S16:
-      copy_values<loco::DataType::S16>(node, cloned);
-      break;
-
-    case loco::DataType::S32:
-      copy_values<loco::DataType::S32>(node, cloned);
-      break;
-
-    case loco::DataType::S64:
-      copy_values<loco::DataType::S64>(node, cloned);
-      break;
-
-    case loco::DataType::BOOL:
-      copy_values<loco::DataType::BOOL>(node, cloned);
-      break;
-
-    default:
-      throw oops::UserExn("Unsupported tensor dtype");
-  }
-
-  return cloned;
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleDiv *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleDiv>();
-  cloned->fusedActivationFunction(node->fusedActivationFunction());
-  return cloned;
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleMean *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleMean>();
-  cloned->keep_dims(node->keep_dims());
-  return cloned;
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleMul *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleMul>();
-  cloned->fusedActivationFunction(node->fusedActivationFunction());
-  return cloned;
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CirclePow *)
-{
-  return _graph->nodes()->create<luci::CirclePow>();
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleRsqrt *)
-{
-  return _graph->nodes()->create<luci::CircleRsqrt>();
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleSqrt *)
-{
-  return _graph->nodes()->create<luci::CircleSqrt>();
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleSquaredDifference *)
-{
-  return _graph->nodes()->create<luci::CircleSquaredDifference>();
-}
-
-luci::CircleNode *CloneNode::visit(const luci::CircleSub *node)
-{
-  auto *cloned = _graph->nodes()->create<luci::CircleSub>();
-  cloned->fusedActivationFunction(node->fusedActivationFunction());
-  return cloned;
-}
-
-} // namespace
 
 namespace luci
 {
